+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog for NeoFS Node
 ### Fixed
 - Control service's Drop call does not clean metabase (#2822)
 - It was impossible to specify memory amount as "1b" (one byte) in config, default was used instead (#2899)
+- Container session token's lifetime was not checked (#2898)
 
 ### Changed
 - neofs-cli allows several objects deletion at a time (#2774)